Clean up parseArgs
